Fix undefined behavior in mpi::datatype::UserDatatype::structured
UserDatatype::structured did not check that the lengths of the component
arrays were equal to `count`, resulting in undefined behavior. I removed
the `count` parameter and matched the other constructors. I also added
assertion messages in the other constructors.

I added a simple test for UserDatatype::structured.

There's some noise in datatype.rs due to formatting that was also fixed
in a separate commit.